Centralize WER report code into evaluate.py, call it from DeepSpeech.py
