+++ requirements.txt
@@ -13,3 +13,4 @@ bs4
 six
 requests
 tables
+attrdict
